bigdataeducation/bigdataeducation/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
jieba==0.42.1
pandas
Werkzeug==3.0.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
openai
gunicorn
pyngrok 